Added structure to app.py, moved individual graph code to different files
